Update mongoose service and add test suite for mongoose service

+++ src/utils/services/mongoose.service.ts
@@ -3,6 +3,7 @@ import mongoose, { Model, QueryOptions, HydratedDocument, FilterQuery, PipelineS
 import { UpdateResult } from 'mongodb';
 import { MongooseConnect } from "../mongoose-connect";
 import { BaseMongooseService } from "../types";
+import { GenericError } from "../errors";
 
 export class MongooseService implements BaseMongooseService {
 
@@ -17,13 +18,41 @@ export class MongooseService implements BaseMongooseService {
     };
 
     private async _reconnectIfDisconnected(): Promise<void> {
+        if (mongoose.connection.readyState === 1) return;
+
         const mongooseConnect = MongooseConnect.getInstance();
         const dbConnectionString = mongooseConnect.dbConnectionString;
 
-        if (mongoose.connection.readyState !== 1) await mongoose.connect(dbConnectionString);
+        await mongoose.connect(dbConnectionString);
     };
 
-    async findById<DocType>(collRef: Model<DocType>, id: string, options?: QueryOptions): Promise<HydratedDocument<DocType> | null> {
+    async save<DocType>(collRef: HydratedDocument<DocType>, options?: SaveOptions | undefined): Promise<DocType | (Document<unknown, any, DocType> & { _id?: unknown; } & Required<{ _id: unknown; }>)> {
+        if (!collRef) throw new GenericError({ error: new Error(`Expected collection reference, but got ${typeof collRef}`), errorCode: 500 });
+
+        await this._reconnectIfDisconnected();
+
+        return collRef.save(options);
+    };
+
+    async insertMany<DocType>(collRef: Model<DocType>, docs: DocType[], options?: InsertManyOptions & { lean: true }): Promise<mongoose.HydratedDocument<mongoose.MergeType<mongoose.MergeType<DocType, DocType>, mongoose.Require_id<DocType>>, {}, {}>[] | MergeType<MergeType<DocType, DocType>, Require_id<DocType>>[]> {
+        if (!collRef) throw new GenericError({ error: new Error(`Expected collection reference, but got ${typeof collRef}`), errorCode: 500 });
+        if (!docs) throw new GenericError({ error: new Error(`Expected array of objects, but got ${typeof docs}`), errorCode: 500 });
+
+        await this._reconnectIfDisconnected();
+
+        type InsertManyResType = Promise<mongoose.HydratedDocument<mongoose.MergeType<mongoose.MergeType<DocType, DocType>, mongoose.Require_id<DocType>>, {}, {}>[] | MergeType<MergeType<DocType, DocType>, Require_id<DocType>>[]>;
+
+        let _query: InsertManyResType = collRef.insertMany(docs);
+        if (options)
+            _query = collRef.insertMany(docs, options);
+
+        return _query;
+    };
+
+    async findById<DocType>(collRef: Model<DocType>, id: string | Types.ObjectId, options?: QueryOptions): Promise<HydratedDocument<DocType> | null> {
+        if (!collRef) throw new GenericError({ error: new Error(`Expected collection reference, but got ${typeof collRef}`), errorCode: 500 });
+        if (!id) throw new GenericError({ error: new Error(`Expected string, but got ${typeof id}`), errorCode: 500 });
+
         await this._reconnectIfDisconnected();
 
         const _query = collRef.findById(id, undefined, options);
@@ -32,6 +61,9 @@ export class MongooseService implements BaseMongooseService {
     };
 
     async findOne<DocType>(collRef: Model<DocType>, query: FilterQuery<DocType>, options?: QueryOptions): Promise<HydratedDocument<DocType> | null> {
+        if (!collRef) throw new GenericError({ error: new Error(`Expected collection reference, but got ${typeof collRef}`), errorCode: 500 });
+        if (!query) throw new GenericError({ error: new Error(`Expected query object, but got ${typeof query}`), errorCode: 500 });
+
         await this._reconnectIfDisconnected();
 
         const _query = collRef.findOne(query, undefined, options);
@@ -40,6 +72,9 @@ export class MongooseService implements BaseMongooseService {
     };
 
     async find<DocType>(collRef: Model<DocType>, query: FilterQuery<DocType>, options?: QueryOptions): Promise<HydratedDocument<DocType>[]> {
+        if (!collRef) throw new GenericError({ error: new Error(`Expected collection reference, but got ${typeof collRef}`), errorCode: 500 });
+        if (!query) throw new GenericError({ error: new Error(`Expected query object, but got ${typeof query}`), errorCode: 500 });
+
         await this._reconnectIfDisconnected();
 
         const _query = collRef.find(query, undefined, options);
@@ -47,15 +82,10 @@ export class MongooseService implements BaseMongooseService {
         return _query;
     };
 
-    async aggregate<DocType>(collRef: Model<DocType>, pipeline?: PipelineStage[] | undefined, options?: AggregateOptions | undefined): Promise<any[]> {
-        await this._reconnectIfDisconnected();
-
-        const _query = collRef.aggregate(pipeline, options);
-
-        return _query;
-    };
-
     async countDocuments<DocType>(collRef: Model<DocType>, query: FilterQuery<DocType>, options?: QueryOptions): Promise<number> {
+        if (!collRef) throw new GenericError({ error: new Error(`Expected collection reference, but got ${typeof collRef}`), errorCode: 500 });
+        if (!query) throw new GenericError({ error: new Error(`Expected query object, but got ${typeof query}`), errorCode: 500 });
+
         await this._reconnectIfDisconnected();
 
         const _query = collRef.countDocuments(query, options);
@@ -63,21 +93,22 @@ export class MongooseService implements BaseMongooseService {
         return _query;
     };
 
-    async save<DocType>(docRef: HydratedDocument<DocType>, options?: SaveOptions | undefined): Promise<DocType | (Document<unknown, any, DocType> & { _id?: unknown; } & Required<{ _id: unknown; }>)> {
+    async aggregate<DocType>(collRef: Model<DocType>, pipeline?: PipelineStage[] | undefined, options?: AggregateOptions | undefined): Promise<any[]> {
+        if (!collRef) throw new GenericError({ error: new Error(`Expected collection reference, but got ${typeof collRef}`), errorCode: 500 });
+        if (!pipeline) throw new GenericError({ error: new Error(`Expected array of objects, but got ${typeof pipeline}`), errorCode: 500 });
+
         await this._reconnectIfDisconnected();
 
-        return docRef.save(options);
-    };
-
-    async insertMany<DocType>(collRef: Model<DocType>, docs: DocType[], options: InsertManyOptions & { lean: true }): Promise<MergeType<MergeType<DocType, DocType>, Require_id<DocType>>[]> {
-        await this._reconnectIfDisconnected();
-
-        const _query = collRef.insertMany(docs, options);
+        const _query = collRef.aggregate(pipeline, options);
 
         return _query;
     };
 
-    async findByIdAndUpdate<DocType>(collRef: Model<DocType>, id: string, updateQuery: UpdateQuery<DocType>, options?: QueryOptions): Promise<HydratedDocument<DocType> | null> {
+    async findByIdAndUpdate<DocType>(collRef: Model<DocType>, id: string | Types.ObjectId, updateQuery: UpdateQuery<DocType>, options?: QueryOptions): Promise<HydratedDocument<DocType> | null> {
+        if (!collRef) throw new GenericError({ error: new Error(`Expected collection reference, but got ${typeof collRef}`), errorCode: 500 });
+        if (!id) throw new GenericError({ error: new Error(`Expected string, but got ${typeof id}`), errorCode: 500 });
+        if (!updateQuery) throw new GenericError({ error: new Error(`Expected update query object, but got ${typeof updateQuery}`), errorCode: 500 });
+
         await this._reconnectIfDisconnected();
 
         const _query = collRef.findByIdAndUpdate(id, updateQuery, options);
@@ -86,6 +117,10 @@ export class MongooseService implements BaseMongooseService {
     };
 
     async findOneAndUpdate<DocType>(collRef: Model<DocType>, query: FilterQuery<DocType>, updateQuery: UpdateQuery<DocType>, options?: QueryOptions): Promise<HydratedDocument<DocType> | null> {
+        if (!collRef) throw new GenericError({ error: new Error(`Expected collection reference, but got ${typeof collRef}`), errorCode: 500 });
+        if (!query) throw new GenericError({ error: new Error(`Expected query object, but got ${typeof query}`), errorCode: 500 });
+        if (!updateQuery) throw new GenericError({ error: new Error(`Expected update query object, but got ${typeof updateQuery}`), errorCode: 500 });
+
         await this._reconnectIfDisconnected();
 
         const _query = collRef.findOneAndUpdate(query, updateQuery, options);
@@ -94,6 +129,10 @@ export class MongooseService implements BaseMongooseService {
     };
 
     async updateOne<DocType>(collRef: Model<DocType>, query: FilterQuery<DocType>, updateQuery: UpdateWithAggregationPipeline | UpdateQuery<DocType>, options?: QueryOptions): Promise<UpdateResult> {
+        if (!collRef) throw new GenericError({ error: new Error(`Expected collection reference, but got ${typeof collRef}`), errorCode: 500 });
+        if (!query) throw new GenericError({ error: new Error(`Expected query object, but got ${typeof query}`), errorCode: 500 });
+        if (!updateQuery) throw new GenericError({ error: new Error(`Expected update query object, but got ${typeof updateQuery}`), errorCode: 500 });
+
         await this._reconnectIfDisconnected();
 
         const _query = collRef.updateOne(query, updateQuery, options);
@@ -102,6 +141,10 @@ export class MongooseService implements BaseMongooseService {
     };
 
     async updateMany<DocType>(collRef: Model<DocType>, query: FilterQuery<DocType>, updateQuery: UpdateWithAggregationPipeline | UpdateQuery<DocType>, options?: QueryOptions): Promise<UpdateResult> {
+        if (!collRef) throw new GenericError({ error: new Error(`Expected collection reference, but got ${typeof collRef}`), errorCode: 500 });
+        if (!query) throw new GenericError({ error: new Error(`Expected query object, but got ${typeof query}`), errorCode: 500 });
+        if (!updateQuery) throw new GenericError({ error: new Error(`Expected update query object, but got ${typeof updateQuery}`), errorCode: 500 });
+
         await this._reconnectIfDisconnected();
 
         const _query = collRef.updateMany(query, updateQuery, options);
@@ -110,6 +153,9 @@ export class MongooseService implements BaseMongooseService {
     };
 
     async findOneAndDelete<DocType>(collRef: Model<DocType>, query: FilterQuery<DocType>, options?: QueryOptions): Promise<HydratedDocument<DocType> | null> {
+        if (!collRef) throw new GenericError({ error: new Error(`Expected collection reference, but got ${typeof collRef}`), errorCode: 500 });
+        if (!query) throw new GenericError({ error: new Error(`Expected query object, but got ${typeof query}`), errorCode: 500 });
+        
         await this._reconnectIfDisconnected();
 
         const _query = collRef.findOneAndDelete(query, options);
